Modulokká bontás
Osztályokat és függvényeket, az "osztalyok" csomagba raktuk.
