+++ Snake.pyproj
@@ -4,7 +4,7 @@
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectGuid>7c7ba3df-1a59-4055-bd7f-6cf38ac6e59c</ProjectGuid>
     <ProjectHome>.</ProjectHome>
-    <StartupFile>pygame_snake.py</StartupFile>
+    <StartupFile>snake.py</StartupFile>
     <SearchPath>
     </SearchPath>
     <WorkingDirectory>.</WorkingDirectory>
@@ -21,10 +21,28 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="osztalyok\kigyo.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="osztalyok\evol.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="osztalyok\globalis.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="osztalyok\main.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="osztalyok\__init__.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="teszt.py">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="pygame_snake.py" />
+    <Compile Include="snake.py" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="osztalyok\" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
   <!-- Uncomment the CoreCompile target to enable the Build command in
