+++ osztalyok/__init__.py
@@ -1,0 +1,4 @@
+from .evol import evol
+from .kigyo import kigyo
+from .globalis import *
+from .main import main
